requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.10
kombu==5.4.2
multidict==6.1.0
passlib==1.7.4
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
redis==5.1.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
twilio==9.3.3
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.14.0
